changed Amount filed type of History Entry to double

+++ src/Lykke.Service.OperationsHistory.Core/Entities/IHistoryLogEntryEntity.cs
@@ -5,7 +5,7 @@ namespace Lykke.Service.OperationsHistory.Core.Entities
     public interface IHistoryLogEntryEntity
     {
         DateTime DateTime { get; set; }
-        decimal Amount { get; set; }
+        double Amount { get; set; }
         string Currency { get; set; }
         string ClientId { get; set; }
         string CustomData { get; set; }
