+++ src/Lykke.Service.OperationsHistory.Core/Entities/IHistoryLogEntryRepository.cs
@@ -6,7 +6,7 @@ namespace Lykke.Service.OperationsHistory.Core.Entities
 {
     public interface IHistoryLogEntryRepository
     {
-        Task AddAsync(DateTime dateTime, decimal amount, string currency, string clientId, string customData, string opType);
+        Task AddAsync(DateTime dateTime, double amount, string currency, string clientId, string customData, string opType);
         Task<IEnumerable<IHistoryLogEntryEntity>> GetAllAsync(string clientId);
     }
 }
